+++ src/questdb/__init__.py
@@ -6,8 +6,12 @@ from typing import Callable as _Callable, Optional as _Optional, Union as _Union
 
 from questdb import _client
 from questdb._client import (
+    Char,
     ConnectionEvent,
     ConnectionEventKind,
+    DateMillis,
+    Geohash,
+    Long256,
     PooledReader,
     PooledSender,
     Protocol,
@@ -33,8 +37,12 @@ from questdb._client import (
 )
 
 __all__ = [
+    'Char',
     'ConnectionEvent',
     'ConnectionEventKind',
+    'DateMillis',
+    'Geohash',
+    'Long256',
     'PooledReader',
     'PooledSender',
     'Protocol',